% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swift-fit.R
\name{swift_fit}
\alias{swift_fit}
\title{Fit TabSwift Tabular Foundation Model}
\usage{
swift_fit(X, y, problem_type = "classification", ...)
}
\arguments{
\item{X}{training data matrix or data.frame of features}

\item{y}{training labels vector}

\item{problem_type}{either "classification" or "regression"}

\item{...}{additional hyperparameters passed down to TabSwift}
}
\value{
A fitted TabSwift model instance
}
\description{
Fit TabSwift Tabular Foundation Model
}
