% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swift-pred.R
\name{swift_predict}
\alias{swift_predict}
\title{Swift Predictions}
\usage{
swift_predict(fit, data, type = "raw", ...)
}
\arguments{
\item{fit}{a fitted model}

\item{data}{a dataset}

\item{type}{'raw'}

\item{...}{for passing parameters to the predict method}
}
\value{
predictions
}
\description{
Swift Predictions
}
